% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disgust.R
\docType{data}
\name{disgust}
\alias{disgust}
\title{Three Domain Disgust Questionnaire (items)}
\format{
A data frame with 20000 rows and 24 variables:
\describe{
\item{id}{Each questionnaire completion's unique ID}
\item{user_id}{Each participant's unique ID}
\item{date}{Date of completion (YYY-mm-dd)}
\item{moral1}{Shoplifting a candy bar from a convenience store}
\item{moral2}{Stealing from a neighbor}
\item{moral3}{A student cheating to get good grades}
\item{moral4}{Deceiving a friend}
\item{moral5}{Forging someone's signature on a legal document}
\item{moral6}{Cutting to the front of a line to purchase the last few tickets to a show}
\item{moral7}{Intentionally lying during a business transaction}
\item{sexual1}{Hearing two strangers having sex}
\item{sexual2}{Performing oral sex}
\item{sexual3}{Watching a pornographic video}
\item{sexual4}{Finding out that someone you don't like has sexual fantasies about you}
\item{sexual5}{Bringing someone you just met back to your room to have sex}
\item{sexual6}{A stranger of the opposite sex intentionally rubbing your thigh in an elevator}
\item{sexual7}{Having anal sex with someone of the opposite sex}
\item{pathogen1}{Stepping on dog poop}
\item{pathogen2}{Sitting next to someone who has red sores on their arm}
\item{pathogen3}{Shaking hands with a stranger who has sweaty palms}
\item{pathogen4}{Seeing some mold on old leftovers in your refrigerator}
\item{pathogen5}{Standing close to a person who has body odor}
\item{pathogen6}{Seeing a cockroach run across the floor}
\item{pathogen7}{Accidentally touching a person's bloody cut}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/disgust.csv}
}
\usage{
disgust
}
\description{
A dataset containing responses to the 21 items in the Three Domain Disgust Questionnaire (Tybur et al.)
}
\keyword{datasets}
