% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{disgust_cors}
\alias{disgust_cors}
\title{Three Domain Disgust Questionnaire (correlations)}
\format{
A data frame with 441 rows and 3 variables:
\describe{
\item{V1}{The first correalted item}
\item{V2}{The second correlated item}
\item{r}{The Pearson's correlation between the first and second item}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/disgust_cors.csv}
}
\usage{
disgust_cors
}
\description{
Correlations among questions on the Three Domain Disgust Questionnaire (Tybur et al.)
}
\keyword{datasets}
