% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_disgust_scores.R
\docType{data}
\name{disgust_scores}
\alias{disgust_scores}
\title{Three Domain Disgust Questionnaire (scores)}
\format{
A data frame with 20000 rows and 6 variables:
\describe{
\item{id}{Each questionnaire completion's unique ID}
\item{user_id}{Each participant's unique ID}
\item{date}{Date of completion (YYY-mm-dd)}
\item{moral}{The mean value for the 7 moral items}
\item{sexual}{The mean value for the 7 sexual items}
\item{pathogen}{The mean value for the 7 pathogen items}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/disgust_scores.csv}
}
\usage{
disgust_scores
}
\description{
A dataset containing subscale scores for to the Three Domain Disgust Questionnaire (Tybur et al.), calculated from \link{disgust}.
}
\keyword{datasets}
