% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_eq_data.R
\docType{data}
\name{eq_data}
\alias{eq_data}
\title{Empathizing Quotient}
\format{
A data frame with 7689 rows and 28 variables:
\describe{
\item{user_id}{Each participant's unique ID}
\item{sex}{The participant's sex}
\item{age}{The participant's age in years}
\item{id}{Each questionnaire completion's unique ID}
\item{starttime}{The time the questionnaire was started (yyyy-MM-dd HH:mm:ss)}
\item{endtime}{The time the questionnaire was completed (yyyy-MM-dd HH:mm:ss)}
\item{Q01}{I can easily tell if someone else wants to enter a conversation.}
\item{Q02}{I really enjoy caring for other people.}
\item{Q03R}{I find it hard to know what to do in a social situation.}
\item{Q04R}{I often find it difficult to judge if something is rude or polite.}
\item{Q05R}{In a conversation, I tend to focus on my own thoughts rather than on what my listener might be thinking.}
\item{Q06}{I can pick up quickly if someone says one thing but means another.}
\item{Q07R}{It is hard for me to see why some things upset people so much.}
\item{Q08}{I find it easy to put myself in somebody else's shoes.}
\item{Q09}{I am good at predicting how someone will feel.}
\item{Q10}{I am quick to spot when someone in a group is feeling awkward or uncomfortable.}
\item{Q11R}{I can't always see why someone should have felt offended by a remark.}
\item{Q12}{I don't tend to find social situations confusing.}
\item{Q13}{Other people tell me I am good at understanding how they are feeling and what they are thinking.}
\item{Q14}{I can easily tell if someone else is interested or bored with what I am saying.}
\item{Q15}{Friends usually talk to me about their problems as they say that I am very understanding.}
\item{Q16}{I can sense if I am intruding, even if the other person doesn't tell me.}
\item{Q17R}{Other people often say that I am insensitive, though I don't always see why.}
\item{Q18}{I can tune into how someone else feels rapidly and intuitively.}
\item{Q19}{I can easily work out what another person might want to talk about.}
\item{Q20}{I can tell if someone is masking their true emotion.}
\item{Q21}{I am good at predicting what someone will do.}
\item{Q22}{I tend to get emotionally involved with a friend's problems.}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/eq_data.csv}
}
\usage{
eq_data
}
\description{
Reverse coded (Q#R) questions coded and strongly disagree = 2, slightly disagree = 1, else = 0. The other questions are coded as strongly agree = 2, slightly agree = 1, else = 0.
}
\details{
Wakabayashi, A., Baron-Cohen, S., Wheelwright, S., Goldenfeld, N., Delaney, J., Fine, D., Smith, R., & Weil, L. (2006). Development of short forms of the Empathy Quotient (EQ-Short) and the Systemizing Quotient (SQ-Short). Personality and Individual Differences, 41(5), 929–940. https://doi.org/10.1016/j.paid.2006.03.017
}
\keyword{datasets}
