% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_family_composition.R
\docType{data}
\name{family_composition}
\alias{family_composition}
\title{Family Composition}
\format{
A data frame with 19169 rows and 11 variables:
\describe{
\item{user_id}{Each participant's unique ID}
\item{sex}{The participant's sex}
\item{age}{The participant's age in years}
\item{momage}{How old was your female parent when you were born?}
\item{dadage}{How old was your male parent when you were born?}
\item{oldbro}{How many older brothers do you have?}
\item{oldsis}{How many older sisters do you have?}
\item{youngbro}{How many younger brothers do you have?}
\item{youngsis}{How many younger sisters do you have?}
\item{twinbro}{How many same-age (e.g. twin) brothers do you have?}
\item{twinsis}{How many same-age (e.g. twin) sisters do you have?}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/family_composition.csv}
}
\usage{
family_composition
}
\description{
Responses to a brief questionnaire about family composition.
}
\keyword{datasets}
