% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getdata}
\alias{getdata}
\title{Create data directory}
\usage{
getdata(dir = "data")
}
\arguments{
\item{dir}{The name of the directory to save the data in (default "data")}
}
\value{
A list of the paths of the new data files
}
\description{
Save the data files used in this class to a directory
}
