% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_matmort.R
\docType{data}
\name{matmort}
\alias{matmort}
\title{Maternal Mortality}
\format{
A data frame with 181 rows and 4 variables:
\describe{
\item{Country}{The full country name}
\item{1990}{Maternal mortality for 1990 (format: "rate \link{lowCI-highCI})}
\item{2000}{Maternal mortality for 2000 (format: "rate \link{lowCI-highCI}")}
\item{2015}{Maternal mortality for 2015 (format: "rate \link{lowCI-highCI}")}
}
}
\source{
\url{https://apps.who.int/gho/data/node.main.15?lang=en}
}
\usage{
matmort
}
\description{
Maternal mortality by country and year from the World Health Organisation.
}
\keyword{datasets}
