% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_mess.R
\docType{data}
\name{mess}
\alias{mess}
\title{Messy Data}
\format{
A data frame with 26 rows and 7 variables:
\describe{
\item{junk}{Junk values (just the word 'junk')}
\item{order}{Meant to be all integer values}
\item{score}{Meant to be all numeric values}
\item{letter}{Meant to be all character values}
\item{good}{Meant to be all logical values}
\item{min_max}{Two integers separated by a dash}
\item{date}{Dates in two different formats}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/mess.csv}
}
\usage{
mess
}
\description{
A dataset with missing values, blank rows, incorrect data types, multiple values in one column, and multiple date types for practicing data import.
}
\keyword{datasets}
