% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_personality_scores.R
\docType{data}
\name{personality_scores}
\alias{personality_scores}
\title{5-Factor Personality Scores}
\format{
A data frame with 15000 rows and 7 variables:
\describe{
\item{user_id}{Each participant's unique ID}
\item{date}{The date this questionnaire was completed}
\item{Ag}{Mean score on Agreeableness items}
\item{Co}{Mean score on Conscientiousness items}
\item{Ex}{Mean score on Extraversion items}
\item{Ne}{Mean score on Neuroticism items}
\item{Op}{Mean score on Openness items}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/personality_scores.csv}
}
\usage{
personality_scores
}
\description{
Archival data from the Face Research Lab of a 5-factor personality questionnaire, with factor score calculated from \link{personality}.
}
\keyword{datasets}
