% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_pets.R
\docType{data}
\name{pets}
\alias{pets}
\title{Pets}
\format{
A data frame with 800 rows and 6 variables:
\describe{
\item{id}{Subject ID}
\item{pet}{Type of pet (dog, cat, ferret)}
\item{country}{What country the pet is from (UK or NL)}
\item{score}{Score on some test (integers around 100)}
\item{age}{Age of the pet in years}
\item{weight}{Weight of the pet in kilograms}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/pets.csv}
}
\usage{
pets
}
\description{
A simulated dataset with one random factor (id), two categorical factors (pet, country) and three continuous variables (score, age, weight). This dataset is useful for practicing plotting.
}
\keyword{datasets}
