% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_psa001_cfd_faces.R
\docType{data}
\name{psa001_cfd_faces}
\alias{psa001_cfd_faces}
\title{Face Characteristics}
\format{
A data frame with 120 rows and 4 variables:
\describe{
\item{Target}{target image - first character designates race, second character designates gender, followed by a unique identifier that matches Target in the CFD dataset}
\item{Race}{target self-identified race (A = asian; B = black; L = latinx; W = white)}
\item{Gender}{target sex (F = female; M = male)}
\item{Age}{Estimate the approximate age of this person (in years) ranges 18.7 to 34.9}
}
}
\source{
\url{https://osf.io/rzgh2/}
}
\usage{
psa001_cfd_faces
}
\description{
Face stimulus characteristics from Psychological Science Accelerator project: To Which World Regions Does the Valence-Dominance Model of Social Perception Apply? \url{https://psyarxiv.com/n26dy}. To be used with \link{psa001_agg}.
}
\details{
Faces from Ma, Correll, & Wittenbrink (2015). The Chicago Face Database: A Free Stimulus Set of Faces and Norming Data. Behavior Research Methods, 47, 1122-1135.

Full data and analysis scripts at \url{https://osf.io/jfwtr/}
}
\keyword{datasets}
