% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_smalldata.R
\docType{data}
\name{smalldata}
\alias{smalldata}
\title{Small Factorial Design: 2w*2b}
\format{
A data frame with 10 rows and 4 variables:
\describe{
\item{id}{Subject ID}
\item{group}{Between-subject factor (control vs experimental group)}
\item{pre}{Score before the intervention}
\item{post}{Score after the intervention}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/smalldata.csv}
}
\usage{
smalldata
}
\description{
Small simulated dataset (n = 5) with one within-subject factor (time) having 2 levels (pre and post) and one beteen-subject factor (group) having two levels(control and experimental). The dataset is in wide format and created with faux.
}
\keyword{datasets}
