% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_sq_data.R
\docType{data}
\name{sq_data}
\alias{sq_data}
\title{Systemizing Quotient}
\format{
A data frame with 4745 rows and 31 variables:
\describe{
\item{user_id}{Each participant's unique ID}
\item{sex}{The participant's sex}
\item{age}{The participant's age in years}
\item{id}{Each questionnaire completion's unique ID}
\item{starttime}{The time the questionnaire was started (yyyy-MM-dd HH:mm:ss)}
\item{endtime}{The time the questionnaire was completed (yyyy-MM-dd HH:mm:ss)}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/sq_data.csv}
}
\usage{
sq_data
}
\description{
Reverse coded (Q#R) questions coded and strongly disagree = 2, slightly disagree = 1, else = 0. The other questions are coded as strongly agree = 2, slightly agree = 1, else = 0.
}
\details{
Wakabayashi, A., Baron-Cohen, S., Wheelwright, S., Goldenfeld, N., Delaney, J., Fine, D., Smith, R., & Weil, L. (2006). Development of short forms of the Empathy Quotient (EQ-Short) and the Systemizing Quotient (SQ-Short). Personality and Individual Differences, 41(5), 929–940. https://doi.org/10.1016/j.paid.2006.03.017
}
\keyword{datasets}
