% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_stroop.R
\docType{data}
\name{stroop}
\alias{stroop}
\title{Stroop Task}
\format{
A data frame with 12500 rows and 5 variables:
\describe{
\item{Country}{The full country name}
\item{1990}{Maternal mortality for 1990 (format: "rate \link{lowCI-highCI})}
\item{2000}{Maternal mortality for 2000 (format: "rate \link{lowCI-highCI}")}
\item{2015}{Maternal mortality for 2015 (format: "rate \link{lowCI-highCI}")}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/stroop.csv}
}
\usage{
stroop
}
\description{
50 simulated subject in a stroop task viewing all combinations of word and ink colours blue, purple, green, red, and brown, 5 times each. Subjects respond with the ink colour. Subjects who do not respond in time have NA for response and rt.
}
\keyword{datasets}
