% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_users.R
\docType{data}
\name{users}
\alias{users}
\title{User Demographics}
\format{
A data frame with 52043 rows and 3 variables:
\describe{
\item{user_id}{Each participant's unique ID}
\item{sex}{The participant's sex}
\item{birthyear}{The participant's year of birth}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/users.csv}
}
\usage{
users
}
\description{
A dataset with unique participant ID, sex and birth year. To be used in conjunction with data from \link{disgust}, \link{disgust_scores}, \link{personality}, \link{personality_scores}, and \link{user}.
}
\keyword{datasets}
