% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_users2.R
\docType{data}
\name{users2}
\alias{users2}
\title{User Demographics 2}
\format{
A data frame with 60000 rows and 3 variables:
\describe{
\item{user_id}{Each participant's unique ID}
\item{birthyear}{The participant's year of birth}
\item{sex}{The participant's sex}
}
}
\source{
\url{https://psyteachr.github.io/msc-data-skills/data/users2.csv}
}
\usage{
users2
}
\description{
A dataset with unique participant ID, birth year, and sex. To be used in conjunction with data from \link{disgust}, \link{disgust_scores}, \link{personality}, \link{personality_scores}, and \link{user}.
}
\keyword{datasets}
