% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{exercise}
\alias{exercise}
\title{Get an exercise}
\usage{
exercise(chapter, filename = NULL, answers = FALSE)
}
\arguments{
\item{chapter}{The chapter that the exercise is for}

\item{filename}{What filename you want to save (defaults to the name of the chapter)}

\item{answers}{Whether or not you want the answers}
}
\value{
Saves a file to the working directory (or path from filename)
}
\description{
Get an exercise
}
\examples{
\dontrun() {
  exercise(1) # get exercise for chapter 1
  exercise(1, answers = TRUE) # get answers for chapter 1
  exercise(1, "exercises/ch1.Rmd") # save into exercises directory
}

}
